+++ L2DNative/L2DNative/RendererL2D.h
@@ -63,7 +63,7 @@ public:
 	void PhysicsSetup(long physicsHandler, float baseLengthM, float airRegistance, float mass);
 	void PhysicsAddSrcParam(long physicsHandler, const char* srcType, const char * paramID, float scale, float weight);
 	void PhysicsAddTargetParam(long phsyicsHandler, const char* targetType, const char* paramID, float scale, float weight);
-	void PhysicsUpdate(long physicsHandler, long hModel, long long time);
+	void PhysicsUpdate(long physicsHandler, long hModel, INT64 time);
 
 	INT64 GetUserTimeMSec();
 	HRESULT BeginRender(long hModel);
